refactor: replace categories with hashtags throughout the application
- Updated router to replace category routes with hashtag routes.
- Modified INews interface to use hashtag_id and hashtag_name instead of category_id and category_name.
- Removed category autocomplete and delete features, added corresponding hashtag features.
- Created new components for editing, creating, and selecting hashtags.
- Updated forms and tables to handle hashtags instead of categories.
- Adjusted validation schemas and payloads to reflect changes from categories to hashtags.
