+++ src/app/constance/navigation.ts
@@ -5,8 +5,8 @@ export const navigation = [
 		secure: false
 	},
 	{
-		link: "/categories",
-		title: "Categories",
+		link: "/hashtags",
+		title: "Hashtags",
 		secure: false
 	},
 	{
