--- src/app/models/categories-model/index.ts
@@ -1,1 +1,0 @@
-export { default as categoriesModel } from "./model.ts";
