+++ src/app/models/hashtags-model/index.ts
@@ -1,0 +1,1 @@
+export { default as hashtagsModel } from "./model.ts";
