+++ src/app/router/view.tsx
@@ -9,9 +9,9 @@ import { NewsPage } from "@/pages/news";
 import { CreateNewsPage } from "@/pages/create-news";
 import { AdminsPage } from "@/pages/admins";
 import { NotFoundPage } from "@/pages/not-found";
-import { CategoriesPage } from "src/pages/categories";
-import { CreateCategoryPage } from "@/pages/create-category";
-import { EditCategoryPage } from "@/pages/edit-category";
+import { HashtagsPage } from "src/pages/hashtags";
+import { CreateHashtagPage } from "@/pages/create-hashtag";
+import { EditHashtagPage } from "@/pages/edit-hashtag";
 import { EditNewsPage } from "@/pages/edit-news";
 import { ReviewNewsPage } from "@/pages/review-news";
 import { ChannelsPage } from "@/pages/channels";
@@ -29,12 +29,12 @@ export const Router = () => {
 							<Route path={"/news/review/:id"} element={<ReviewNewsPage />} />
 							<Route path={"/create-news"} element={<CreateNewsPage />} />
 							<Route path={"/admins"} element={<AdminsPage />} />
-							<Route path={"/categories"} element={<CategoriesPage />} />
+							<Route path={"/hashtags"} element={<HashtagsPage />} />
 							<Route
-								path={"/create-category"}
-								element={<CreateCategoryPage />}
+								path={"/create-hashtag"}
+								element={<CreateHashtagPage />}
 							/>
-							<Route path={"/category/:id"} element={<EditCategoryPage />} />
+							<Route path={"/hashtag/:id"} element={<EditHashtagPage />} />
 							<Route path={"/channels"} element={<ChannelsPage />} />
 						</Route>
 					</Route>
