--- src/features/category-autocomplete/index.ts
@@ -1,1 +1,0 @@
-export { CategoryAutocomplete } from "./view";
