--- src/features/delete-category/index.ts
@@ -1,1 +1,0 @@
-export { DeleteCategory } from "./view.tsx";
