+++ src/features/delete-hashtag/index.ts
@@ -1,0 +1,1 @@
+export { DeleteHashtag } from "./view.tsx";
