--- src/features/edit-category/index.ts
@@ -1,1 +1,0 @@
-export { EditCategory } from "./view.tsx";
