+++ src/features/edit-hashtag/index.ts
@@ -1,0 +1,1 @@
+export { EditHashtag } from "./view.tsx";
