+++ src/features/hashtag-autocomplete/index.ts
@@ -1,0 +1,1 @@
+export { HashtagAutocomplete } from "./view";
