--- src/features/select-category/index.ts
@@ -1,1 +1,0 @@
-export { SelectCategory } from "./view";
