+++ src/features/select-hashtag/index.ts
@@ -1,0 +1,1 @@
+export { SelectHashtag } from "./view";
