--- src/pages/categories/index.ts
@@ -1,1 +1,0 @@
-export { CategoriesPage } from "./view";
