+++ src/pages/create-hashtag/index.ts
@@ -1,0 +1,1 @@
+export { CreateHashtagPage } from "./view";
