--- src/pages/edit-category/index.ts
@@ -1,1 +1,0 @@
-export { EditCategoryPage } from "./view.tsx";
