+++ src/pages/edit-hashtag/index.ts
@@ -1,0 +1,1 @@
+export { EditHashtagPage } from "./view.tsx";
