+++ src/pages/hashtags/index.ts
@@ -1,0 +1,1 @@
+export { HashtagsPage } from "./view";
