--- src/widgets/categories-table/index.ts
@@ -1,1 +1,0 @@
-export { CategoriesTable } from "./view.tsx";
