--- src/widgets/category-form/index.ts
@@ -1,1 +1,0 @@
-export { CategoryForm } from "./view.tsx";
