+++ src/widgets/hashtag-form/index.ts
@@ -1,0 +1,1 @@
+export { HashtagForm } from "./view.tsx";
