+++ src/widgets/hashtags-table/index.ts
@@ -1,0 +1,1 @@
+export { HashtagsTable } from "./view.tsx";
