+++ src/widgets/news-form/model.ts
@@ -36,16 +36,16 @@ export const schema = z.object({
 		.max(100, "Max 100 characters for ad link")
 		.optional()
 		.nullable(),
-	category_id: z.string().optional(),
-	category_name: z
+	hashtag_id: z.string().optional(),
+	hashtag_name: z
 		.string()
-		.nonempty("Category is required")
+		.nonempty("Hashtag is required")
 		.refine(
 			val => {
-				// Может быть либо числом (ID), либо валидным именем категории
+				// Может быть либо числом (ID), либо валидным названием хэштега
 				return /^\d+$/.test(val) || /^[a-z0-9_]{2,50}$/.test(val);
 			},
-			"Invalid category"
+			"Invalid hashtag"
 		),
 	poster_link: z.string().min(1, "Poster image is required"),
 	translations: z.array(newsFormSchema)
